Simplify DandelionConfig
