+++ p2p/src/lib.rs
@@ -53,6 +53,6 @@ pub use peer::Peer;
 pub use peers::Peers;
 pub use serv::{DummyAdapter, Server};
 pub use store::{PeerData, State};
-pub use types::{Capabilities, ChainAdapter, DandelionConfig, Direction, Error, P2PConfig,
+pub use types::{Capabilities, ChainAdapter, Direction, Error, P2PConfig,
                 PeerInfo, ReasonForBan, TxHashSetRead, MAX_BLOCK_HEADERS, MAX_LOCATORS,
                 MAX_PEER_ADDRS};
