+++ servers/src/grin/seed.rs
@@ -29,6 +29,7 @@ use time::{self, now_utc};
 use hyper;
 
 use p2p;
+use pool::DandelionConfig;
 use util::LOGGER;
 
 const SEEDS_URL: &'static str = "http://grin-tech.org/seeds.txt";
@@ -42,6 +43,7 @@ const DNS_SEEDS: &'static [&'static str] = &[
 pub fn connect_and_monitor(
 	p2p_server: Arc<p2p::Server>,
 	capabilities: p2p::Capabilities,
+	dandelion_config: DandelionConfig,
 	seed_list: Box<Fn() -> Vec<SocketAddr> + Send>,
 	stop: Arc<AtomicBool>,
 ) {
@@ -73,7 +75,7 @@ pub fn connect_and_monitor(
 						tx.clone(),
 					);
 
-					update_dandelion_relay(peers.clone(), p2p_server.dandelion_config.clone());
+					update_dandelion_relay(peers.clone(), dandelion_config.clone());
 
 					prev = current_time;
 				}
@@ -159,7 +161,7 @@ fn monitor_peers(
 	}
 }
 
-fn update_dandelion_relay(peers: Arc<p2p::Peers>, dandelion_config: p2p::DandelionConfig) {
+fn update_dandelion_relay(peers: Arc<p2p::Peers>, dandelion_config: DandelionConfig) {
 	// Dandelion Relay Updater
 	let dandelion_relay = peers.get_dandelion_relay();
 	if dandelion_relay.is_empty() {
